Allow require bundler/gem_tasks to fail if bundler is not found
When including as a gem in another project bundler won't be loaded during a bundle install.
In turn this means that although running gem install redsnow would work, attempting to add to
another project's gemfile and then running bundle install would not.
This allows the LoadError to be ignored and show a warning to stout so installation can continue.